fix(rating): cast level->numeric in backfill+decay (live level col is double precision; round(float8,int) doesn't exist)

+++ supabase/changes/2026-07-02_rating_engine_v2.sql
@@ -58,7 +58,7 @@ begin
          where h.profile_id = p.id and h.match_id is not null), 0);
 
     update public.profiles p set
-      rating = round(coalesce(level, public.level_from_elo(coalesce(elo, 1000))), 2),
+      rating = round(coalesce(level, public.level_from_elo(coalesce(elo, 1000)))::numeric, 2),
       last_competitive_match_at = (
         select max(h.created_at) from public.ranking_history h
          where h.profile_id = p.id and h.match_id is not null);
@@ -314,7 +314,7 @@ begin
 
   -- (b) gentle rating decay after 60 days idle
   for r in
-    select p.id, coalesce(p.rating, coalesce(p.level, 0)) as rating
+    select p.id, coalesce(p.rating, coalesce(p.level, 0))::numeric as rating
       from profiles p
      where coalesce(p.is_admin, false) = false
        and coalesce(p.competitive_matches, 0) >= 5
